move: 패키지 이동 /config -> /config/security
